requirements.txt
pydantic
google-generativeai
lancedb
pyarrow
rdkit-pypi
mdanalysis
tomli